Change VerifiedBuild type to full int

+++ HeaderParser.cs
@@ -111,7 +111,7 @@ namespace WowClientDB2MySQLTableGenerator
 
             structure.Members.Add(new CStructureMemberInfo()
             {
-                TypeName = "int16",
+                TypeName = "int32",
                 Name = "VerifiedBuild"
             });
